Added config file feature

+++ src/commands/Bne.cc
@@ -6,8 +6,8 @@
 using namespace std;
 
 void Bne::run(MipsClient & client, const LabelTable & table) const {
-    int32_t s = client.getRegister(tokens.at(1).valAsNumber());
-    int32_t t = client.getRegister(tokens.at(3).valAsNumber());
+    int64_t s = client.getRegister(tokens.at(1).valAsNumber());
+    int64_t t = client.getRegister(tokens.at(3).valAsNumber());
     if (s == t) return;
 
     if (tokens.at(5).getType() == Token::INT) {
